+++ ux/inc/_html.navBar.php
@@ -46,13 +46,6 @@
                         </li>
                         <li class="nav-item nav-dropdown">
                             <a class="nav-link nav-dropdown-toggle" href="#"><i class="fa fa-life-ring"></i> Body</a>
-	                        <?php /*if ($isDeveloper) { */?><!--
-                            <ul class="nav-dropdown-items">
-                                <li class="nav-item">
-                                    <a class="nav-link" href="body/body.html"><i class="fa fa-line-chart"></i> Body <span class="badge badge-danger">WIP</span></a>
-                                </li>
-                            </ul>
-	                        --><?php /*}*/ ?>
                             <ul class="nav-dropdown-items">
                                 <li class="nav-item">
                                     <a class="nav-link" href="body/weight.html"><i class="fa fa-heart"></i> Body Weight</a>
@@ -78,8 +71,20 @@
                         <li class="nav-item">
                             <a class="nav-link" href="food/food.html"><i class="fa fa-cutlery""></i> Food</a>
                         </li>
-                        <li class="nav-item">
-                            <a class="nav-link" href="nomie/nomie.html"><i class="fa fa-street-view"></i> Nomie</a>
+                        <li class="nav-item nav-dropdown">
+                            <a class="nav-link nav-dropdown-toggle" href="#"><i class="fa fa-street-view"></i> Nomie</a>
+                            <ul class="nav-dropdown-items">
+                                <li class="nav-item">
+                                    <a class="nav-link" href="nomie/nomie.html"><i class="fa fa-dashboard"></i>
+                                        Dashboard</a>
+                                </li>
+                            </ul>
+                            <ul class="nav-dropdown-items">
+                                <li class="nav-item">
+                                    <a class="nav-link" href="nomie/trackers.html"><i class="fa fa-check-square-o"></i>
+                                        Trackers</a>
+                                </li>
+                            </ul>
                         </li>
 
                         <li class="nav-title">